distributed api.py over several files. Added typechecking dependent import on resource\tag.py to avoid circular refs
