+++ modeltestSDK/api/__init__.py
@@ -1,0 +1,6 @@
+from .campaign import CampaignAPI
+from .test import TestAPI, FloaterTestAPI, WindCalibrationAPI, WaveCalibrationAPI
+from .sensor import SensorAPI
+from .timeseries import TimeseriesAPI
+from .tag import TagAPI
+from .floater_config import FloaterConfigAPI
